pyproject.toml
[tool.poetry]
name = "mnist-chest-xray-classification"
version = "0.1.0"
description = "This project implements a deep learning solution for binary classification of chest X-rays using the ChestMNIST dataset from MedMNIST"
authors = ["Piotr Gryko <piotr.gryko@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
medmnist = "^3.0.2"
jupyter = "^1.1.1"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
notebook = "^7.3.2"
mlflow = "^2.19.0"
black = {extras = ["jupyter"], version = "^24.10.0"}
structlog = "^24.4.0"
python-decouple = "^3.8"
albumentations = "^1.4.24"
torchinfo = "^1.8.0"
pynvml = "^12.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
ruff = "^0.8.4"
pylama = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pylama]
max_line_length = 200